refactor: 설문 결과 DTO inner class로 변경
