+++ survwey/survwey/src/main/java/mcnc/survwey/api/survey/response/service/SurveyReplyService.java
@@ -3,44 +3,29 @@ package mcnc.survwey.api.survey.response.service;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import mcnc.survwey.domain.question.enums.QuestionType;
 import mcnc.survwey.domain.objAnswer.ObjAnswer;
 import mcnc.survwey.domain.objAnswer.repository.ObjAnswerRepository;
 import mcnc.survwey.domain.objAnswer.service.ObjAnswerService;
 import mcnc.survwey.domain.question.Question;
-import mcnc.survwey.api.survey.response.dto.answered.AnsweredQuestionDTO;
-import mcnc.survwey.api.survey.response.dto.result.QuestionResultDTO;
-import mcnc.survwey.api.survey.response.dto.result.SurveyResultQueryDTO;
-import mcnc.survwey.domain.question.repository.QuestionRepository;
 import mcnc.survwey.domain.respond.Respond;
 import mcnc.survwey.api.survey.response.dto.reply.ReplyDTO;
 import mcnc.survwey.domain.respond.repository.RespondRepository;
 import mcnc.survwey.domain.respond.service.RespondService;
-import mcnc.survwey.api.survey.response.dto.result.SelectionResultDTO;
 import mcnc.survwey.domain.subjAnswer.SubjAnswer;
 import mcnc.survwey.domain.subjAnswer.repository.SubjAnswerRepository;
 import mcnc.survwey.domain.subjAnswer.service.SubjAnswerService;
 import mcnc.survwey.domain.survey.Survey;
-import mcnc.survwey.domain.survey.repository.SurveyRepository;
 import mcnc.survwey.domain.survey.service.SurveyService;
-import mcnc.survwey.api.survey.response.dto.answered.AnsweredSurveyDTO;
-import mcnc.survwey.api.survey.response.dto.result.SurveyResultDTO;
 import mcnc.survwey.api.survey.response.dto.reply.SurveyReplyDTO;
 import mcnc.survwey.domain.user.User;
-import mcnc.survwey.api.survey.response.dto.result.AgeCountDTO;
-import mcnc.survwey.api.survey.response.dto.result.GenderCountDTO;
 import mcnc.survwey.domain.user.service.UserService;
 import mcnc.survwey.global.exception.custom.CustomException;
 import mcnc.survwey.global.exception.custom.ErrorCode;
-import org.springframework.cache.annotation.Cacheable;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
 import java.util.*;
 import java.util.stream.Collectors;
-
-import static mcnc.survwey.domain.subjAnswer.QSubjAnswer.subjAnswer;
 
 @Service
 @Slf4j
@@ -111,8 +96,6 @@ public class SurveyReplyService {
             throw new CustomException(HttpStatus.BAD_REQUEST, ErrorCode.QUESTION_NOT_MATCH_TO_SURVEY);
         }
     }
-
-
 
 
 }
