+++ survwey/survwey/src/main/java/mcnc/survwey/domain/user/service/UserService.java
@@ -2,10 +2,9 @@ package mcnc.survwey.domain.user.service;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import mcnc.survwey.api.survey.response.dto.result.SurveyResultDTO;
 import mcnc.survwey.domain.user.enums.Gender;
 import mcnc.survwey.domain.user.User;
-import mcnc.survwey.api.survey.response.dto.result.AgeCountDTO;
-import mcnc.survwey.api.survey.response.dto.result.GenderCountDTO;
 import mcnc.survwey.domain.user.repository.UserRepository;
 import mcnc.survwey.global.exception.custom.CustomException;
 import mcnc.survwey.global.exception.custom.ErrorCode;
@@ -40,15 +39,15 @@ public class UserService {
      * @param surveyId
      * @return
      */
-    public List<GenderCountDTO> getGenderCountListBySurveyId(Long surveyId) {
+    public List<SurveyResultDTO.GenderCountDTO> getGenderCountListBySurveyId(Long surveyId) {
         List<Object[]> recordList = userRepository.findGenderCountBySurveyId(surveyId);
 
         //응답 Map 생성
-        Map<Gender, GenderCountDTO> genderCountDTOMap = new LinkedHashMap<>();
+        Map<Gender, SurveyResultDTO.GenderCountDTO> genderCountDTOMap = new LinkedHashMap<>();
 
         //Key 초기화
-        genderCountDTOMap.put(Gender.M, new GenderCountDTO(Gender.M.getValue(), 0));
-        genderCountDTOMap.put(Gender.F, new GenderCountDTO(Gender.F.getValue(), 0));
+        genderCountDTOMap.put(Gender.M, new SurveyResultDTO.GenderCountDTO(Gender.M.getValue(), 0));
+        genderCountDTOMap.put(Gender.F, new SurveyResultDTO.GenderCountDTO(Gender.F.getValue(), 0));
 
         //응답 Map에 값 할당
         for (Object[] record : recordList) {
@@ -66,7 +65,7 @@ public class UserService {
      * @param surveyId
      * @return
      */
-    public List<AgeCountDTO> getAgeGroupCountBySurveyId(Long surveyId) {
+    public List<SurveyResultDTO.AgeCountDTO> getAgeGroupCountBySurveyId(Long surveyId) {
         List<LocalDate> birthList = userRepository.findBirthBySurveyId(surveyId);
         //DB 조회 결과를 나이대(10대, 20대 ...)를 key로 하는 Map 생성
         Map<Integer, Integer> ageMap = groupAgesByDecade(birthList);
@@ -105,14 +104,14 @@ public class UserService {
      * @param ageMap
      * @return
      */
-    private List<AgeCountDTO> mapAgeGroupsToDTO(Map<Integer, Integer> ageMap) {
-        List<AgeCountDTO> ageCountDTOList = new ArrayList<>();
-        ageCountDTOList.add(new AgeCountDTO("10대 미만", ageMap.getOrDefault(0, 0)));
+    private List<SurveyResultDTO.AgeCountDTO> mapAgeGroupsToDTO(Map<Integer, Integer> ageMap) {
+        List<SurveyResultDTO.AgeCountDTO> ageCountDTOList = new ArrayList<>();
+        ageCountDTOList.add(new SurveyResultDTO.AgeCountDTO("10대 미만", ageMap.getOrDefault(0, 0)));
         for (int decade = 1; decade <= 7; decade++) {
             String ageGroup = decade * 10 + "대";
-            ageCountDTOList.add(new AgeCountDTO(ageGroup, ageMap.getOrDefault(decade, 0)));
+            ageCountDTOList.add(new SurveyResultDTO.AgeCountDTO(ageGroup, ageMap.getOrDefault(decade, 0)));
         }
-        ageCountDTOList.add(new AgeCountDTO("80세 이상", ageMap.getOrDefault(8, 0)));
+        ageCountDTOList.add(new SurveyResultDTO.AgeCountDTO("80세 이상", ageMap.getOrDefault(8, 0)));
         return ageCountDTOList;
     }
